instastore/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-browser-reload==1.21.0
django-cors-headers==4.6.0
django-filter==24.3
django-htmx==1.27.0
django-prometheus==2.4.1
django-tailwind==4.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
prometheus_client==0.23.1
psycopg2-binary==2.9.10
pytailwindcss
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
